requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.7
django-bootstrap4==22.3
django-stdimage==6.0.1
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.4.0
PyMySQL==1.0.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0